[master]活动审核活动渠道、活动奖品处理逻辑

+++ nas-common/src/main/java/com/minyan/nascommon/po/ActivityChannelPO.java
@@ -19,7 +19,7 @@ public class ActivityChannelPO implements Serializable {
   private Long id;
 
   /** 活动id */
-  private Long activityId;
+  private Integer activityId;
 
   /** 渠道名称 */
   private String channelName;
